refactor: 클라이언트 컴포넌트 버튼을 배너 이미지에서 분리 (#185)
